Made a simple user frontpage
Added 5 different 'hyperlinks' for now, which do not work yet with some basic animation as basis of the layout. In the future the colors will be changed, they are just there to identify different quadrants, and many other smaller things, like the white buttons around the dropdown menu options etc. Moreover, I want to add an animated gradient as the background to everything, which is what I will be working on for now if there are no other more important things.